requirements.txt
ipykernel
jupyter==1.0.*
pandas==2.*
boto3==1.*
botocore
toml
# pdf2image==1.17.*
# amazon-textract-textractor
langchain==0.3.*
langchain-openai==0.2.*
langchain-community==0.3.*
langchain-aws==0.2.*
langchain-ollama==0.2.*
langchain-experimental==0.3.* 
langgraph
neo4j==5.*
pydantic==2.10.1
pdfplumber
chainlit
pymupdf 
pillow
duckduckgo-search
numexpr
streamlit